Decrease SlurmdTimeout to 48s from 300s
If the NFS server gets preempted, it's up and running well within 5 minutes,
since we poll it every 60 seconds. This means that Slurm never recognizes that
it went down, so any jobs running on it will be marked as FAIL, not NODE_FAIL,
and will not be requeued.

Given the 60 second polling (and overhead from shutdown/startup times),
48 seconds should give enough time for Slurm to realize that the NFS is down
and requeue its jobs.