Docker tasks on preempted nodes requeue automatically
The shutdown signal kills often kills systemd processes first, before killing
users' processes. This means that dockerd receives a kill signal before the
container itself, which invokes a clean shutdown of the container (which the
Slurm controller is aware of) before the controller is aware that the node has
failed.

This adds a custom shutdown script to systemd that signals to the controller that
this node has failed. It is guaranteed to run before dockerd is terminated.